+++ NEWS.md
@@ -171,6 +171,8 @@ Standard library changes
 Deprecated or removed
 ---------------------
 
+* `map` and `foreach` lost their single-argument methods ([#52631]).
+
 
 External dependencies
 ---------------------